Fix compiler warnings

+++ src/CryptoNoteCore/CryptoNoteFormatUtils.cpp
@@ -535,7 +535,7 @@ bool get_block_longhash(cn_context &context, const Block& b, Hash& res) {
   if (b.majorVersion >= BLOCK_MAJOR_VERSION_5) {
     Crypto::Hash hash_1, hash_2;
     cn_fast_hash(bd.data(), bd.size(), hash_1);
-    Crypto::pump[hash_1.data[0] & 3](bd.data(), hash_2, bd.size(), hash_1.data, sizeof(hash_1));
+    Crypto::pump[hash_1.data[0] & 3](bd.data(), hash_2, (int)bd.size(), hash_1.data, sizeof(hash_1));
 
     res = hash_2;
   }
